preprocessRequirement cleanup
